+++ src/utils.ts
@@ -6,7 +6,9 @@ const comparisons = {
   '<' : (a:number, b: number) => a < b,
   '>' : (a:number, b: number) => a > b,
   '<=': (a:number, b: number) => a <= b,
-  '>=': (a:number, b: number) => a >= b
+  '>=': (a:number, b: number) => a >= b,
+  '!=': (a:number, b: number) => a != b,
+  // '!empty': (a:number, b: number) => !!a,
 };
 
 /**