+++ src/base64.ts
@@ -1,7 +1,10 @@
+
 /**
- * Decodes Base64 url encoded strings.
- * @param {string} str
- * @returns {string} Decoded string.
+ * Decodes a base64 encoded string.
+ *
+ * @param {string} str - The base64 encoded string to decode.
+ * @param {boolean} urlSafe - A boolean indicating if the input string is URL safe. Defaults to true.
+ * @returns {string} The decoded string in UTF-8 format.
  */
 function b64Decode(str: string, urlSafe = true): string {
   if (urlSafe)
@@ -10,11 +13,13 @@ function b64Decode(str: string, urlSafe = true): string {
 	return Buffer.from(str + pad(str), "base64").toString("utf8");
 }
 
+
 /**
- * Encodes the given data and returns it as a url encoded base64 string.
- *
- * @param {string} data - The data to be encrypted.
- * @return {string} The encrypted data in base64 format.
+ * Encodes a given string into Base64 format.
+ * 
+ * @param {string} str - The string to be encoded.
+ * @param {boolean} urlSafe - Optional boolean to determine if the output should be URL safe. Defaults to true.
+ * @returns {string} The Base64 encoded string. If `urlSafe` is true, the output will be modified to be URL safe.
  */
 function b64Encode(str: string, urlSafe = true): string {
   let b64 = Buffer.from(str).toString("base64");