+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 - Add isNull method
 - Add isUndefined method
 - Add "from" parameter to method isIn()
+- Add "!=" to comparisons
 - Update typings of most methods for easier use with Typescript
 - Typings are more strict. Resulting in possible breaking changes from Checkard version 2.x.x.
-- other methods than primitives and non-primitives methods do not check for input type anymore. their output is now Boolean only. Resulting in possible breaking changes from Checkard version 2.x.x.
+- Other methods than primitives and non-primitives methods do not check for input type anymore. their output is now Boolean only. Resulting in possible breaking changes from Checkard version 2.x.x.
 - Improve Documentation 
 
 # 2.30.0 (Jan 29th 2025)